+++ Assets/Plugins/GameplayAbilitySystem/GameplayEffects/Components/GameplayEffectAttributeEntityComponent.cs
@@ -1,0 +1,10 @@
+using System;
+using Unity.Entities;
+namespace GameplayAbilitySystem.GameplayEffects.Systems {
+    [Serializable]
+    public struct GameplayEffectAttributeEntityComponent : IComponentData {
+        public Entity Value;
+        public static implicit operator Entity(GameplayEffectAttributeEntityComponent e) { return e.Value; }
+        public static implicit operator GameplayEffectAttributeEntityComponent(Entity e) { return new GameplayEffectAttributeEntityComponent { Value = e }; }
+    }
+}
