+++ Assets/Plugins/GameplayAbilitySystem/GameplayEffects/Components/GameplayEffectTargetComponent.cs
@@ -1,0 +1,10 @@
+using System;
+using Unity.Entities;
+namespace GameplayAbilitySystem.GameplayEffects.Systems {
+    [Serializable]
+    public struct GameplayEffectTargetComponent : IComponentData {
+        public Entity Value;
+        public static implicit operator Entity(GameplayEffectTargetComponent e) { return e.Value; }
+        public static implicit operator GameplayEffectTargetComponent(Entity e) { return new GameplayEffectTargetComponent { Value = e }; }
+    }
+}
