+++ Assets/Plugins/GameplayAbilitySystem/GameplayEffects/Components/GameplayEffectTargetSCDComponent.cs
@@ -1,0 +1,12 @@
+﻿using Unity.Entities;
+
+namespace GameplayAbilitySystem.GameplayEffects.Components {
+    public struct GameplayEffectTargetSCDComponent : ISharedComponentData {
+        public Entity Value;
+        public static implicit operator Entity(GameplayEffectTargetSCDComponent e) { return e.Value; }
+        public static implicit operator GameplayEffectTargetSCDComponent(Entity e) { return new GameplayEffectTargetSCDComponent { Value = e }; }
+    }
+ 
+}
+
+
